% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmponline.R
\name{retrieve_date}
\alias{retrieve_date}
\title{Function to retrieve plans created from a certain date}
\usage{
retrieve_date(date = Sys.Date(), ..., instance = "tudelft", verbose = T)
}
\arguments{
\item{date}{retrieve plans created from this date. defaults to today.}

\item{...}{pass through arguments to \code{dmponline_auth()}}

\item{instance}{string. Whether you are using the default DMPonline (dmponline.dcc.ac.uk) or your own organisation has an instance.}

\item{verbose}{logical. Whether or not to print the API request (default = \code{TRUE}).}
}
\description{
Function to retrieve plans created from a certain date
}
