% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmponline.R
\name{retrieve_dmp}
\alias{retrieve_dmp}
\title{Function to retrieve a DMP based on the plan number}
\source{
https://github.com/DMPRoadmap/roadmap/wiki/API-documentation
}
\usage{
retrieve_dmp(dmp_number, full_plan = F, ..., instance = "tudelft")
}
\arguments{
\item{dmp_number}{number of the plan (i.e. https://dmponline.dcc.ac.uk/plans/<plan_number>)}

\item{full_plan}{logical. Whether or not to download the full plan (only available in API v0).}

\item{...}{pass through arguments to \code{dmponline_auth()}}
}
\description{
Function to retrieve a DMP based on the plan number
}
\details{
Takes the DMP number as an argument and returns the DMP as a list.
Default arguments returns the DMP metadata, including contact details of owner,
dates of creation and last updated, and project description.
The full plan also contains all the questions (with answers).
}
