% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmponline.R
\name{retrieve_heartbeat}
\alias{retrieve_heartbeat}
\title{Function to ping the DMPonline API.}
\usage{
retrieve_heartbeat()
}
\description{
Basic function to ping the DMPonline API. No authentication required.
If the call returns Status: 200, then the API was successfully reached.
}
