% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{retrieve_statistics}
\alias{retrieve_statistics}
\title{Function to retrieve a DMPs based on the users in 3mE}
\usage{
retrieve_statistics(type = "plans", start_date = NULL, end_date = NULL)
}
\arguments{
\item{type}{string with type of request. Can be \code{users_joined}, \code{using_template}, \code{plans_by_template}, or \code{plans}.}

\item{start_date}{start date to retrieve stats in form YYYY-MM-DD.}

\item{end_date}{end date to retrieve stats in form YYYY-MM-DD.}
}
\description{
Function to retrieve a DMPs based on the users in 3mE
}
