% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmponline.R
\name{retrieve_date}
\alias{retrieve_date}
\title{Function to retrieve plans created on a certain date}
\usage{
retrieve_date(date = Sys.Date(), ...)
}
\arguments{
\item{date}{retrieve plans created on this date. defaults to today.}

\item{...}{pass through arguments to \code{dmponline_auth()}}
}
\description{
Function to retrieve plans created on a certain date
}
