% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{retrieve_statistics}
\alias{retrieve_statistics}
\title{Function to retrieve statistics for DMPonline usage}
\usage{
retrieve_statistics(
  type = "plans",
  start_date = NULL,
  end_date = NULL,
  instance = "tudelft",
  ...
)
}
\arguments{
\item{type}{string with type of request. Can be \code{users_joined}, \code{using_template}, \code{plans_by_template}, or \code{plans}.}

\item{start_date}{start date as string from when to retrieve stats in form "YYYY-MM-DD".}

\item{end_date}{end date as string until when to retrieve stats in form "YYYY-MM-DD".}

\item{instance}{string. Whether you are using the default DMPonline (dmponline.dcc.ac.uk) or your own organisation has an instance.}

\item{...}{pass through arguments to \code{dmponline_auth()}}
}
\description{
Function to retrieve statistics for DMPonline usage
}
\examples{
# Retrieve statistics for plans created between 1st January 2023 and 31st January 2023 (inclusive)
retrieve_statistics(type = "plans", start_date = "2023-01-01", end_date = "2023-01-31")
}
